// src/app/models/dealer.model.ts
export class Dealer{
    id: string;
    name: string;
    contactNumber: number;
    faxNumber: string;
    code: string;
    country: string;
    type: string;
    houseNo: string;
    state: string;
    city: string;
    emailId: string;
    address1: string;
    address2: string;
    zipCode: string;
    createdBy: string;
    loginRequired: string;
    oldName: string;
    countryCode: string;
}